debug some errors and add comment to each function

+++ backend/convert.js
@@ -1,11 +1,9 @@
 'use strict';
 
 exports.convertAudio = function(req, res) {
-    console.log(123333);
     var fs = require('fs-extra');
     var toolName = req.params.toolname;
     var inputType =req.params.inputtype;
-    console.log(toolName);
     var audioName;
     var audioFolder;
     var audioPath;
@@ -21,34 +19,39 @@ exports.convertAudio = function(req, res) {
         default:
             break;
     }
-    console.log(audioPath);
-    switch (toolName){
-        case "Sphinx-4":
-            var exec = require('child_process').exec;
-            var cmd1 = 'cd '+audioFolder;
-            var cmd2 = 'sox '+audioName+' -c 1 -r 16000 0-converted.wav';
-            exec(cmd1+' ; '+cmd2, function(error, stdout, stderr) {
-                console.log('convert ok');
-                fs.unlinkSync(audioFolder + '/' + (fs.readdirSync(audioFolder))[1]);
-                res.json({
-                    convertMsg: 'Convert ok!!',
+    if (audioName !== "0-converted.wav" && audioName !== undefined){
+        switch (toolName){
+            case "Sphinx-4":
+                var exec = require('child_process').exec;
+                var cmd1 = 'cd '+audioFolder;
+                var cmd2 = 'sox '+audioName+' -c 1 -r 16000 -b 16 0-converted.wav';
+                exec(cmd1+' ; '+cmd2, function(error, stdout, stderr) {
+                    console.log('convert ok');
+                    fs.unlinkSync(audioFolder + '/' + (fs.readdirSync(audioFolder))[1]);
+                    res.json({
+                        convertMsg: 'Convert ok!!',
+                    });
                 });
-            });
-        break;
-        case "Kaldi":
-            var exec = require('child_process').exec;
-            var cmd1 = 'cd '+audioFolder;
-            var cmd2 = 'sox '+audioName+' -c 1 -r 8000 0-converted.wav';
-            console.log(cmd1+' ; '+cmd2);
-            exec(cmd1+' ; '+cmd2, function(error, stdout, stderr) {
-                console.log('convert ok');
-                fs.unlinkSync(audioFolder + '/' + (fs.readdirSync(audioFolder))[1]);
-                res.json({
-                    convertMsg:'Convert ok!!'
+            break;
+            case "Kaldi":
+                var exec = require('child_process').exec;
+                var cmd1 = 'cd '+audioFolder;
+                var cmd2 = 'sox '+audioName+' -c 1 -r 8000 -b 16 0-converted.wav';
+                console.log(cmd1+' ; '+cmd2);
+                exec(cmd1+' ; '+cmd2, function(error, stdout, stderr) {
+                    console.log('convert ok');
+                    fs.unlinkSync(audioFolder + '/' + (fs.readdirSync(audioFolder))[1]);
+                    res.json({
+                        convertMsg:'Convert ok!!'
+                    });
                 });
-            });
-            break;
-        default:
-            break;
-    };  
+                break;
+            default:
+                break;
+        };
+    }
+    else 
+        res.json({
+            convertMsg:'Your audio is missing'
+        });
 };
